fix errors

+++ src/openssl/signatures.c
@@ -657,7 +657,7 @@ xmlSecOpenSSLSignatureDsaSign(xmlSecOpenSSLSignatureCtxPtr ctx, xmlSecBufferPtr 
     if (dsaSignBuf == NULL) {
         xmlSecInternalError2("xmlSecBufferCreate",
                              NULL,
-                             "size=%lu", dsaSignBufSize);
+                             "size=%lu", (unsigned long)dsaSignBufSize);
         goto done;
     }
     ret = EVP_PKEY_sign(pKeyCtx, xmlSecBufferGetData(dsaSignBuf), &dsaSignBufSize, ctx->dgst, ctx->dgstSize);
@@ -1145,7 +1145,7 @@ xmlSecOpenSSLSignatureEcdsaSign(xmlSecOpenSSLSignatureCtxPtr ctx, xmlSecBufferPt
     if (ecSignBuf == NULL) {
         xmlSecInternalError2("xmlSecBufferCreate",
             NULL,
-            "size=%lu", ecSignBufSize);
+            "size=%lu", (unsigned long)ecSignBufSize);
         goto done;
     }
     ret = EVP_PKEY_sign(pKeyCtx, xmlSecBufferGetData(ecSignBuf), &ecSignBufSize, ctx->dgst, ctx->dgstSize);
